QCLI Test automation - merged context feature test into one
